Bug fixes and edit/insert improvements. Brick insert changed to be extension of brick edit (dashboard, elements, tool, etc). Fixed some preexisting bugs with saving attributes according to their form value. AttributeKeyCategoryItem now has a saveAttributeForm() method that behaves alot like saveAttribute(), and the respective methods were also updated in AttributeKey.
Set new convention for attribute type validateForm returns: NULL means the form value is nullable and the attribute can be cleared, TRUE means it is ready to save, and FALSE or a string means there is an error (string is the error message).

Added "handles" to the validation helper, which basically just lets you attach a handle to an error message for reference, and then check if that specific error exists with $error->has('your_handle'). This is used on the AKCI attribute form for displaying contextual errors with attributes.

Fixed bug on search results with persistent jsInitArgs.
Fixed bug on search results with column config not automatically refreshing the form. Still want to redo the column config tool js to bring it up to speed with the other js.