Implementation of CONCURRENT_SESSION_LIMIT.
The following definitions have been added to the configuration (defaults
have been set in concrete/config/base.php)

   ENABLE_CONCURRENT_SESSION_LIMIT		true|false
   CONCURRENT_SESSION_LIFE				number (8)	[hours sessions keep]
   CONCURRENT_SESSION_LIMIT				number (1)  [allowed sessions]

Session Limit can be over-ridden on a per user basis by adding the
'allowed_sessions' UserAttribute to the user and setting it to the desired
number.   setting this attribute to -1 will allow unlimited sessions.

The user defined as USER_SUPER_ID automatically is allowed unlimited
sessions.  This user cannot be limited.

Sessions are determined by PHP Session ID, IP Address and Browser Version.
All must match for a session to continue.

A new database table is required for the sessions.  Use the "refresh
database" functionality under Dashboard -> Sitewide Settings -> Refresh
Schema and ensure the 'core database tables' option is ticked.

There is currently no way to remove or view sessions from the Dashboard
interface.  This may be added in the near future.