Small Image block add/edit tweaks - minimal styling
New link location (page on current site, external, etc...) is within a header tag, making the text within look big, black and kindof ugly. Certainly out of keeping. Probably not a great idea to have the select field within the H2 here, but hey...it works for now. I've just moved the font-size (the em default sizing is thwarted by the fixed 18px pixel font-size on the parent H2 element) inline, and spaced it away from the H2 text slightly. Also taken the opportunity to consolidate the table structure and spacing on the max width and height inputs at the bottom -- adding cellpadding and removed the non-breaking spaces. Again, not perfect, but consistent and swaps bad build practice for ever-so-slightly better bad build practice.