Prevent non-returning events from prematurely triggering the fire() return
By adding a quick is_null check when each event is fired, we should be able to help mitigate legitimate and desirable event getting ignored. Setting the priority so that your returning event fires last is the surest bet to make it behave, but this small tweak should ensure that if only one triggered method of several returns content, its priority shouldn't matter as it's the only one that'll trigger a return at all.