Making Countries and Provinces lists maintainable
Re-worked pull requests 591 & 592 that I made on the master branch into
the 553 devel branch.
https://github.com/concrete5/concrete5/pull/592
https://github.com/concrete5/concrete5/pull/591

The countries and state/provinces lists are now searched for in the
Config table; if not there it is set from class data which is now the
default. This enables other software to potentially manipulate/modify
the list.

As the Concrete5_Helper_Lists_... do not appear to be implemented yet
for
these helpers, I have updated both the Concrete and Concrete/Core
copies,
though one set of updates will no doubt be redundant.