Provides events that top and tail an area and block wrapper
Re-worked pull request 594 that I made on the master branch into
the 553 devel branch. Also includes tidied ordering of log that had
already been accepted into the master branch (because I needed it to
check this was behaving properly)
https://github.com/concrete5/concrete5/pull/594
With these events, an addon will be able to place wrappers round an area
or blocks in an area without modifying the theme. This will facilitate
add-ons that decorate areas or blocks within areas or restructure a
page. As a simple example, an add-on could force an area to render
hidden and then show it later, thus overcoming 'Flash of unstyled
content'. I also have in mind many more involved possibilities involving
restructuring an area as it renders.