Add "Block Wrapper Template" functionality to Areas
This adds the ability for themes to have little "block wrapper templates" for an area. See https://github.com/jordanlev/c5_block_wrapper_templates for a detailed explanation of how to use this feature.
Note that this functionality replaces the "block wrapper tokens" that I added in 5.5.1(?)... because it provides the same thing but in a more elegant way (plus lets you do a lot more than the tokens did).