Added customizable pattern when using meta title When we're not using the meta title for pages, concrete5 allows to format how the page name is rendered as a title (via the `PAGE_TITLE_FORMAT` define). When we're using the meta title, this meta title is used directly for the title tag. Anyway someone could want to use a pattern also for the meta-titled pages to have a standard title across the site, for example to have always a title like 'SiteName :: PageTitle'. So I introduced the `PAGE_METATITLE_FORMAT` define, which by default print out only the meta-title, but can be customized to be the same as the `PAGE_TITLE_FORMAT` define.
Furthermore the decision to use the meta-title or the page name (in the header_required.php) is made only with an instruction like `if($metaTitle) {...}`.
But this fails when the meta-title is an empty string (someone may want to have an empty title... maybe) or is "0" (zero), since in php "" == false and also "" == false.