Dashboard keywords alignment for install & update
When updating to 5.6.0 the dashboard pages are saved localized with the
t function, whereas when installing the keywords are saved in English.
This pull request make 4 changes:
- keywords are saved always in English
- aligns the keywords of the install and the update (they were a bit
  different)
- Allows the usage of translated keywords during Intelligent Search
- Normalizes the dashboard.xml indentation

As s side note: the keywords must be extracted from the dashboard.xml
file and processed by the translation system (Transifex at the moment).
PS: my i18n.php tool has been updated to do that: see
https://github.com/mlocati/concrete5-build