Add 'country' field to Form block
The form block is missing a commonly used field: the country.
This pull request adds a new field named 'country'. For i18n purposes
the selected country is saved with its code: its name is shown
depending on the current locale.

I furthermore fixed a few things in the Excel export function:
* the indentation was messed up
* moved the document title from the header 'Content-Title' to the
  html <title> tag
* added some t() function
* table headers are in <thead>, body cells in <tbody>
* added some missing Loader::helper('text')->entities()