auto embedding of pictures in css
it probably needs more testing but I thought I'd create that pull
request anyway to see what others think about this.

5.6.1 has a lot of great performance improvements but I thought about
other ways to improve the performance and came up with an idea to
reduce the number of http request.

I'm doing this by including pictures in the css. Simply set
EMBED_CSS_IMAGES to true in site.php and all your pictures are
inline in the css.