Fix Form on Global Block Design Edit Dialog
Fix form on Gllobal Block design edit dialog.

Current values were not being displayed in form field when editing Block
styles for global areas.

This was due to the fact that the `CollectionBlockStyles` table was
being queries for the Global Area name, but needed to be queries for the
Stack Area name.

Checks if Collection Object is `STACKS_PAGE_TYPE` and if so sets Area
Handle to `STACKS_AREA_NAME`