Fix Display in Multiselect to use html `<br />`
Fix display in multiselect where `<br />` tags are being displayed as
text.

Taking a pramatic approach and handling select attributes only. This may
not cover all cases, but in general difficult to determine in which
cases we want to escape text and which cases we want it to be html