Fix Mutliple Page Delete in Dashboard Page Search
Fix multiple page delete in dashboard page search.

See commit ada9071c76684e2880138e232c1c62cd0b9418e7

In this commit the default was removed. Possible fix for this issue may
instead be to reinstate the parameter default of 'ACTIVE' to accommodate
other code in the wild that assumed a default