Fix Mutliple Page Delete in Dashboard Page Search
Fix multiple page delete in dashboard page search.

See commit 1df01e2d34442c1089bc728e08f216c21d5201e9

In this commit the default was removed. Possible fix for this issue may
instead be to reinstate the parameter default of 'ACTIVE' to accommodate
other code in the wild that assumed a default