Fixed a bug in 'getOrCreate'
Fixed a bug in 'getOrCreate' when using an exact subset of an existing group combination.

If a group combination is created with 3 or more groups for a page and then that, or another, page is created/updated with an exact subset of the original/existing group combination, the getOrCreate function's SQL statement continues to target the original group combination making it impossible to actually create a group combination with a subset of that original/existing group combination.

Eg. Given a group combination of Group1+Group2+Group3, it is impossible to create a group combination of Group1+Group2 because the SQL statement keeps matching the group combination for Group1+Group2+Group3 as it contains Group1+Group2.