Use en_US number format to solve 'priority' bug
Issue: If the administrator has its interface language set to 'Dutch', commas are being used in decimals. This results in an issue for Google, as it expects the field 'prority' with a decimal point. (see https://support.google.com/webmasters/answer/183668?hl=en)

Possible solution: set the locale for numbers (back) to American English. 

I've tested this and it works.

Proposing this for the 5.6 branch because it is a bug.