Inhibit creation of another admin user
When creating/updating an user, we check if it does not have the 'admin'
username.
Let's perform this check in a case-insensitive way.

BTW, this shouldn't be a big issue, since we already check if another user
with the new username already exists.
So, we could also remove these checks...