Removed even number element restriction
I removed the even number of elements restriction AND changed the default number of pages over to 9 to be more similar to the current number of starting elements.

There is still a restriction where you cannot have less than 7 elements showing. I think if someone did want to do something like this it would be more of a [Previous][$current][Next] sort of look but I'm not going to take the time to implement that here.