Additional check for the Auto Nav block
The getNavigationArray() method in the autonav block controller is responsible for preparing the array of nav items. This commit, in the beginning of that function, checks if the parent page is excluded from nav or if it has been set to exclude child pages from nav and only then lets the execution of the method continue.