Updated Auto-Nav block controller to minimize database queries
The getNavigationArray method is responsible for populating the array of navigation items. Later, in getNavItems method, the controller goes through each navigation item and excludes the pages based on certain conditions. This approach is not good as it fetches all the pages and for each page separate database queries are made (i.e for fetching page object using Page::getByID). Hence, there is a possibility that the menu on front end displays only 10 items, but in the back end more than 2000 queries are being made just for that instance of Auto-Nav block.

This commit adds additional checks in the beginning of getNavigationArray to prevent the fetching of unnecessary pages.